renderer: Split up common stuffs to main source
