+++ src/emulator/renderer/include/renderer/texture_cache_state.h
@@ -5,9 +5,13 @@
 #include <gxm/types.h>
 #include <psp2/gxm.h>
 
+#include <cstdint>
 #include <array>
+#include <functional>
 
-namespace renderer::gl {
+struct MemState;
+
+namespace renderer {
 constexpr size_t TextureCacheSize = 1024;
 typedef uint64_t TextureCacheTimestamp;
 typedef uint32_t TextureCacheHash;
@@ -15,6 +19,9 @@ typedef uint32_t TextureCacheHash;
 typedef std::array<emu::SceGxmTexture, TextureCacheSize> TextureCacheGxmTextures;
 typedef std::array<TextureCacheTimestamp, TextureCacheSize> TextureCacheTimestamps;
 typedef std::array<TextureCacheHash, TextureCacheSize> TextureCacheHashes;
+typedef std::function<void(std::size_t)> TextureCacheStateSelectCallback;
+typedef std::function<void(std::size_t, const void*)> TextureCacheStateConfigureTextureCallback;
+typedef std::function<void(std::size_t, const void*, const MemState&)> TextureCacheStateUploadTextureCallback;
 
 struct TextureCacheState {
     size_t used = 0;
@@ -22,6 +29,8 @@ struct TextureCacheState {
     TextureCacheGxmTextures gxm_textures;
     TextureCacheTimestamps timestamps;
     TextureCacheHashes hashes;
-    GLObjectArray<TextureCacheSize> textures;
+    TextureCacheStateSelectCallback select_callback;
+    TextureCacheStateConfigureTextureCallback configure_texture_callback;
+    TextureCacheStateUploadTextureCallback upload_texture_callback;
 };
 } // namespace renderer
