+++ src/emulator/renderer/src/gl/texture_formats.cpp
@@ -548,73 +548,5 @@ GLenum translate_minmag_filter(SceGxmTextureFilter src) {
     }
 }
 
-size_t bits_per_pixel(SceGxmTextureBaseFormat base_format) {
-    switch (base_format) {
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U8:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S8:
-        return 8;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U4U4U4U4:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U8U3U3U2:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U1U5U5U5:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U5U6U5:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S5S5U6:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U8U8:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S8S8:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U16:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S16:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_F16:
-        return 16;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U8U8U8U8:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S8S8S8S8:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U2U10U10U10:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U16U16:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S16S16:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_F16F16:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_F32:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_F32M:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_X8S8S8U8:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_X8U24:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U32:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S32:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_SE5M9M9M9:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_F11F11F10:
-        return 32;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_F16F16F16F16:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U16U16U16U16:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S16S16S16S16:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_F32F32:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U32U32:
-        return 64;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_PVRT2BPP:
-        return 2;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_PVRT4BPP:
-        return 4;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_PVRTII2BPP:
-        return 2;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_PVRTII4BPP:
-        return 4;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_UBC1:
-        return 4;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_UBC2:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_UBC3:
-        return 8;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_YUV420P2:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_YUV420P3:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_YUV422:
-        return 16; // NOTE: I'm not sure this is right.
-    case SCE_GXM_TEXTURE_BASE_FORMAT_P4:
-        return 4;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_P8:
-        return 8;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U8U8U8:
-    case SCE_GXM_TEXTURE_BASE_FORMAT_S8S8S8:
-        return 24;
-    case SCE_GXM_TEXTURE_BASE_FORMAT_U2F10F10F10:
-        return 32;
-    }
-
-    return 0;
-}
-
 } // namespace texture
 } // namespace renderer
