+++ src/emulator/renderer/src/texture_palette.cpp
@@ -1,12 +1,11 @@
 #include <renderer/functions.h>
-
-#include "profile.h"
+#include <renderer/profile.h>
 
 #include <mem/ptr.h>
 #include <util/log.h>
 #include <gxm/types.h>
 
-namespace renderer::gl {
+namespace renderer {
 namespace texture {
 
 void palette_texture_to_rgba_4(uint32_t *dst, const uint8_t *src, size_t width, size_t height, const uint32_t *palette) {
